+++ ServerGUI/Login/CryptoConfiguration.cs
@@ -1,0 +1,8 @@
+﻿namespace Login
+{
+    public class CryptoConfiguration
+    {
+        public byte[] Key { get; set; }
+        public byte[] IV { get; set; }
+    }
+}
