+++ ServerGUI/Login/Services/ILoginService.cs
@@ -1,0 +1,9 @@
+﻿namespace Login.Services
+{
+    public interface ILoginService
+    {
+        bool RegisterAccount(string data);
+        bool CheckData(string data);
+        bool ChangePassword(string data);
+    }
+}
