--- ServerGUI/LoginLibrary/CryptoConfiguration.cs
@@ -1,8 +1,0 @@
-﻿namespace LoginLibrary
-{
-    public class CryptoConfiguration
-    {
-        public byte[] Key { get; set; }
-        public byte[] IV { get; set; }
-    }
-}
