--- ServerGUI/LoginLibrary/Services/ILoginService.cs
@@ -1,9 +1,0 @@
-﻿namespace LoginLibrary.Services
-{
-    public interface ILoginService
-    {
-        bool RegisterAccount(string data);
-        bool CheckData(string data);
-        bool ChangePassword(string data);
-    }
-}
