+++ ServerGUI/Server/Events/ServerEvents.cs
@@ -1,0 +1,19 @@
+﻿using Prism.Events;
+
+namespace Server.Events
+{
+    public class ServerStartedEvent : PubSubEvent
+    {
+
+    }
+
+    public class WrongServerConfigurationEvent : PubSubEvent<string>
+    {
+
+    }
+
+    public class ServerLogsChanged : PubSubEvent<string>
+    {
+
+    }
+}
