+++ ServerGUI/Server/Events/UserEvents.cs
@@ -1,0 +1,14 @@
+﻿using Prism.Events;
+
+namespace Server.Events
+{
+    public class UserLoggedInEvent : PubSubEvent
+    {
+
+    }
+
+    public class UserDisconnectedEvent : PubSubEvent
+    {
+
+    }
+}
