+++ ServerGUI/Server/ServerConfiguration.cs
@@ -1,0 +1,10 @@
+﻿namespace Server
+{
+    public class ServerConfiguration
+    {
+        public int Port { get; set; }
+        public int WeatherBufferSize { get; set; }
+        public int LoginBufferSize { get; set; }
+        public string IpAddress { get; set; }
+    }
+}
