+++ ServerGUI/Server/Services/IServerService.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace Server.Services
+{
+    public interface IServerService
+    {
+        Task StartServer(string ipAddress, int port);    
+    }
+}
