--- ServerGUI/ServerLibrary/Events/UserEvents.cs
@@ -1,14 +1,0 @@
-﻿using Prism.Events;
-
-namespace ServerLibrary.Events
-{
-    public class UserLoggedInEvent : PubSubEvent
-    {
-
-    }
-
-    public class UserDisconnectedEvent : PubSubEvent
-    {
-
-    }
-}
