--- ServerGUI/ServerLibrary/ServerConfiguration.cs
@@ -1,10 +1,0 @@
-﻿namespace ServerLibrary
-{
-    public class ServerConfiguration
-    {
-        public int Port { get; set; }
-        public int WeatherBufferSize { get; set; }
-        public int LoginBufferSize { get; set; }
-        public string IpAddress { get; set; }
-    }
-}
