--- ServerGUI/ServerLibrary/Services/IServerService.cs
@@ -1,9 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace ServerLibrary.Services
-{
-    public interface IServerService
-    {
-        Task StartServer(string ipAddress, int port);    
-    }
-}
