+++ ServerGUI/Storage.DAL/Services/IStorageService.cs
@@ -1,4 +1,4 @@
-﻿using StorageLibrary.Models;
+﻿using Storage.Models;
 using System.Collections.Generic;
 using System.Threading.Tasks;
