+++ ServerGUI/Storage/Context/IUserDataContext.cs
@@ -1,0 +1,16 @@
+﻿using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Infrastructure;
+using Storage.Models;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Storage.Context
+{
+    public interface IUserDataContext
+    {
+        DbSet<UserData> UserDatas { get; set; }
+        DatabaseFacade DatabaseFacade { get; }
+
+        Task<int> SaveChangesAsync(CancellationToken cancellationToken = default);
+    }
+}
