+++ ServerGUI/Storage/DatabaseConfiguration.cs
@@ -1,0 +1,7 @@
+﻿namespace Storage
+{
+    public class DatabaseConfiguration
+    {
+        public string ConnectionString { get; set; }
+    }
+}
