+++ ServerGUI/Storage/Models/UserData.cs
@@ -1,0 +1,9 @@
+﻿namespace Storage.Models
+{
+    public class UserData
+    {
+        public long Id { get; set; }
+        public string Usernane { get; set; }
+        public string Password { get; set; }
+    }
+}
