--- ServerGUI/StorageLibrary/Context/IUserDataContext.cs
@@ -1,16 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Infrastructure;
-using StorageLibrary.Models;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace StorageLibrary.Context
-{
-    public interface IUserDataContext
-    {
-        DbSet<UserData> UserDatas { get; set; }
-        DatabaseFacade DatabaseFacade { get; }
-
-        Task<int> SaveChangesAsync(CancellationToken cancellationToken = default);
-    }
-}
