--- ServerGUI/StorageLibrary/DatabaseConfiguration.cs
@@ -1,7 +1,0 @@
-﻿namespace StorageLibrary
-{
-    public class DatabaseConfiguration
-    {
-        public string ConnectionString { get; set; }
-    }
-}
