--- ServerGUI/StorageLibrary/Models/UserData.cs
@@ -1,9 +1,0 @@
-﻿namespace StorageLibrary.Models
-{
-    public class UserData
-    {
-        public long Id { get; set; }
-        public string Usernane { get; set; }
-        public string Password { get; set; }
-    }
-}
