+++ ServerGUI/Weather/Services/IWeatherService.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace Weather.Services
+{
+    public interface IWeatherService
+    {
+        Task<string> GetWeather(string location, int days);
+    }
+}
