+++ ServerGUI/Weather/WeatherApiConfiguration.cs
@@ -1,0 +1,7 @@
+﻿namespace Weather
+{
+    public class WeatherApiConfiguration
+    {
+        public string ApiKey { get; set; }
+    }
+}
