--- ServerGUI/WeatherLibrary/Services/IWeatherService.cs
@@ -1,9 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace WeatherLibrary.Services
-{
-    public interface IWeatherService
-    {
-        Task<string> GetWeather(string location, int days);
-    }
-}
