--- ServerGUI/WeatherLibrary/WeatherApiConfiguration.cs
@@ -1,7 +1,0 @@
-﻿namespace WeatherLibrary
-{
-    public class WeatherApiConfiguration
-    {
-        public string ApiKey { get; set; }
-    }
-}
